webapp/services/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

# HTTP client
httpx==0.25.2

# Redis
redis==5.0.1

# Rate limiting
slowapi==0.1.9

# Celery for background tasks
celery==5.3.4

# Coordinate transformation
pyproj==3.6.1

# WebSocket support
websockets==12.0
